feat: Set initial theme mode in InitialThemeState
This commit sets the initial theme mode in the InitialThemeState class. If the initialThemeMode parameter is provided, it will be used as the theme mode. Otherwise, the default value of ThemeMode.light will be used. This ensures that the theme mode is always set, even if no initial theme mode is specified.

The colorPalette and appTextStyle properties are also updated based on the theme mode, using the LightModeColorPalette and LightAppTextStyles classes for the light theme mode, and the DarkModeColorPalette and DarkAppTextStyles classes for the dark theme mode.

These changes improve the consistency and reliability of the theme state initialization.